requirements.txt
numpy==1.26.0
pandas==1.3.4
jupyter==1.0.0
seaborn==0.11.2
ipykernel==6.4.2
scikit-learn==1.3.1
umap-learn
pyod